+++ tezt/lib_tezos/sc_rollup_node.mli
@@ -118,6 +118,10 @@ val string_of_history_mode : history_mode -> string
     GC, and the [gc_frequency] is [1] by default to make it runs on
     every occasion during tests.
 
+    [remote_signer] makes the rollup node use a remote signer at the provided
+    URL to sign messages for keys that were registered with the ["remote:"]
+    scheme.
+
 *)
 val create :
   ?runner:Runner.t ->
@@ -126,6 +130,7 @@ val create :
   ?color:Log.Color.t ->
   ?data_dir:string ->
   base_dir:string ->
+  ?remote_signer:Uri.t ->
   ?event_pipe:string ->
   ?metrics_addr:string ->
   ?metrics_port:int ->
@@ -151,6 +156,7 @@ val create_with_endpoint :
   ?color:Log.Color.t ->
   ?data_dir:string ->
   base_dir:string ->
+  ?remote_signer:Uri.t ->
   ?event_pipe:string ->
   ?metrics_addr:string ->
   ?metrics_port:int ->